Refactor order of helpers.py imports, reorganize some zip utils, refactor gdalutils.convert.
